Move parseHorizons function into config module
